[git] Set `core.pager` to `less -x1,5`
This sets tabs to rendered as 4 spaces by `git` commands. With the
options `-x1,5`, the `1` is to skip over the leading column in diff
outputs (that contains `+/-/space`). The first tab stop is then set to
`5`, with subsequent tab stops continuing with the same spacing as the
previous values (i.e. `5 - 1` gives a spacing of `4`).